feat: cleanups and moving static functions to helpers module
