+++ notebooks/arb_dao_grant_distribution/output/dao_grant_2023-11-30_2023-12-14.json
@@ -49,7 +49,7 @@
             },
             "contractInputsValues": {
                 "gaugeAddresses": "[0xcf9f895296F5e1D66a7D4dcf1d92e1B435E9f999,0x40F6bE0b26B229C4C3b2C3344e91f8eEe22a62fF,0x04fc019017eD3F921D5ec11fFf84B870744BA0d1,0x159be31493C26F8F924b3A2a7F428C2f41247e83,0x2eB5661002b68EBE887d29d415c3A3b52536912C,0xeF767E740D83d410794519c2F93Db32e44359a5C,0x96056DF5a61B758005a219a052D72085C45df92c,0x972539E9d340a915775C004715f286a166F067Fd,0xcF853F14EF6111435Cb39c0C43C66366cc6300F1,0xd956246EA5b06DEa930F0A7feC1FFf000436e3f2,0x089a2E069700B0f2B16f69A6e02b01818ce22e7d,0x011417BBED6FC9cefF36C032D431b0eFcBA7f8B3,0xD7a5122a4e88EC9fBE94301A0Bff6114ddFe5d4B,0x82d2c7B67Eaa5028c89BE86CeA8e1DF5bd2119A1,0x260cbb867359a1084eC97de4157d06ca74e89415,0xfC745035F31BCbaEb2D1a89aA9171495c671F6cE,0xa14453084318277b11d38FbE05D857A4f647442B,0xd6B875d62c2661eaB66472F36c672e4B512f1135,0xc3522B0A58d9f12f33E542FC3c3C051b4AF01c83,0x34BDBEcd3B911a8DCB17aC68B961a8D16db52673,0x138E37c3885169DB38e046D5c814C0e95566566c]",
-                "amountsPerPeriod": "[10285739999999999781721,4088996781926967742037,4627346730289855436,2057148000000000138243,8783098674253618810325,5508348782297970501531,1156598280085974920439,4152907019850138567562,8228592000000000552972,9892601850312881651916,4236353977819977444596,1784320658464075449955,3699966147785832163208,3379171287534291423071,5012995613671129831345,7559659329472728131804,6784056476854422726319,8466853840494899486657,412958850935387772551,1737405081509403999007,5625000000000000000000]",
+                "amountsPerPeriod": "[10285739999999999781721,4089526527789112151367,4642516923028935593,2057148000000000138243,8795028252683170649106,5515429425656517196330,1154736361173951536329,4146622159649094101041,8228592000000000552972,9877286955132145521929,4229911590662886737845,1784612010134654383364,3694460550273898206796,3375773478138003611093,5023234950690144614782,7562932045837046644010,6787426767400665994500,8472484573097988686640,412391273385329043321,1734420561372312249659,5625000000000000000000]",
                 "maxPeriods": "[2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2]"
             }
         },
